xfig_new_text accepts rotate=angle and eps2eps qualifiers
If a rotate=angle qualifier is given, the text is wrapped
in a  \rotatebox{angle}{ }  command from the graphicx package
and the eps2eps qualifier is set in order to find a large enough Bounding Box.

If the eps2eps qualifier is set, eps2eps is run on the autoepsfile.

[Original Date:   Sun Feb 28 18:20:12 2010 +0100]